refactor(api): restructure filters definitions into smaller files
