+++ src/lib/constants.js
@@ -1,0 +1,5 @@
+export const ATOMIC = "atomic"
+export const UNSTYLED = "unstyled"
+
+export const IMAGE = "IMAGE"
+export const HORIZONTAL_RULE = "HORIZONTAL_RULE"
