+++ src/lib/filters/entities.test.js
@@ -1,355 +1,9 @@
-import {
-  EditorState,
-  convertFromHTML,
-  ContentState,
-  convertFromRaw,
-} from "draft-js"
+import { EditorState, convertFromRaw } from "draft-js"
 
-import {
-  preserveAtomicBlocks,
-  resetBlockDepth,
-  resetBlockType,
-  filterInlineStyle,
-  resetAtomicBlocks,
-  filterEntityType,
-  filterEditorState,
-} from "./normalize"
+import { IMAGE, HORIZONTAL_RULE } from "../constants"
+import { resetAtomicBlocks, filterEntityType } from "./entities"
 
-const UNSTYLED = "unstyled"
-const ATOMIC = "atomic"
-const IMAGE = "IMAGE"
-const HORIZONTAL_RULE = "HORIZONTAL_RULE"
-
-const BOLD = "BOLD"
-
-const UNORDERED_LIST_ITEM = "unordered-list-item"
-const HEADER_TWO = "header-two"
-const HEADER_FIVE = "header-five"
-
-describe("normalize", () => {
-  describe("#preserveAtomicBlocks", () => {
-    it("works", () => {
-      const contentState = convertFromRaw({
-        entityMap: {
-          "4": {
-            type: "IMAGE",
-            mutability: "IMMUTABLE",
-            data: {
-              src: "../static/example-lowres-image.jpg",
-            },
-          },
-          "5": {
-            type: "EMBED",
-            mutability: "IMMUTABLE",
-            data: {
-              url: "http://www.youtube.com/watch?v=y8Kyi0WNg40",
-            },
-          },
-        },
-        blocks: [
-          {
-            key: "d3071",
-            text: " ",
-            type: "unstyled",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [
-              {
-                offset: 0,
-                length: 1,
-                key: 4,
-              },
-            ],
-            data: {},
-          },
-          {
-            key: "affm4",
-            text: " ",
-            type: "unstyled",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [
-              {
-                offset: 0,
-                length: 1,
-                key: 5,
-              },
-            ],
-            data: {},
-          },
-          {
-            key: "abbm4",
-            text: " ",
-            type: "unstyled",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-        ],
-      })
-
-      expect(
-        preserveAtomicBlocks(EditorState.createWithContent(contentState), [
-          IMAGE,
-        ])
-          .getCurrentContent()
-          .getBlockMap()
-          .map((b) => b.getType())
-          .toJS(),
-      ).toEqual({
-        d3071: "atomic",
-        affm4: "unstyled",
-        abbm4: "unstyled",
-      })
-    })
-
-    it("no normalisation = no change", () => {
-      const contentState = convertFromRaw({
-        entityMap: {
-          "4": {
-            type: "IMAGE",
-            mutability: "IMMUTABLE",
-            data: {
-              src: "../static/example-lowres-image.jpg",
-            },
-          },
-        },
-        blocks: [
-          {
-            key: "d3071",
-            text: " ",
-            type: "unstyled",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [
-              {
-                offset: 0,
-                length: 1,
-                key: 4,
-              },
-            ],
-            data: {},
-          },
-        ],
-      })
-
-      const editorState = EditorState.createWithContent(contentState)
-
-      expect(preserveAtomicBlocks(editorState, [])).toBe(editorState)
-    })
-  })
-
-  describe("#resetBlockDepth", () => {
-    it("normalises depth to a given number", () => {
-      const contentState = convertFromRaw({
-        entityMap: {},
-        blocks: [
-          {
-            key: "d3071",
-            text: "Depth 0",
-            type: "unordered-list-item",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-          {
-            key: "affm4",
-            text: "Depth 1",
-            type: "unordered-list-item",
-            depth: 1,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-          {
-            key: "abbm4",
-            text: "Depth 2",
-            type: "unordered-list-item",
-            depth: 2,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-        ],
-      })
-      const editorState = resetBlockDepth(
-        EditorState.createWithContent(contentState),
-        1,
-      )
-      expect(
-        editorState
-          .getCurrentContent()
-          .getBlockMap()
-          .map((block) => block.getDepth())
-          .toJS(),
-      ).toEqual({ abbm4: 1, affm4: 1, d3071: 0 })
-    })
-
-    it("no normalisation = no change", () => {
-      const editorState = EditorState.createWithContent(
-        ContentState.createFromBlockArray(
-          convertFromHTML(`<ul><li>Depth 0</li></ul>`),
-        ),
-      )
-      expect(resetBlockDepth(editorState, 1)).toBe(editorState)
-    })
-  })
-
-  describe("#resetBlockType", () => {
-    it("works", () => {
-      const contentState = convertFromRaw({
-        entityMap: {},
-        blocks: [
-          {
-            key: "d3071",
-            text: "P",
-            type: "unstyled",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-          {
-            key: "affm4",
-            text: "UL",
-            type: "unordered-list-item",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-          {
-            key: "abbm1",
-            text: "H1",
-            type: "header-one",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-          {
-            key: "abbm4",
-            text: "H2",
-            type: "header-two",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-          {
-            key: "abbm8",
-            text: "H3",
-            type: "header-three",
-            depth: 0,
-            inlineStyleRanges: [],
-            entityRanges: [],
-            data: {},
-          },
-        ],
-      })
-      const editorState = resetBlockType(
-        EditorState.createWithContent(contentState),
-        [UNORDERED_LIST_ITEM, HEADER_TWO],
-      )
-      expect(
-        editorState
-          .getCurrentContent()
-          .getBlockMap()
-          .map((block) => block.getType())
-          .toJS(),
-      ).toEqual({
-        abbm1: "unstyled",
-        abbm4: "header-two",
-        abbm8: "unstyled",
-        affm4: "unordered-list-item",
-        d3071: "unstyled",
-      })
-    })
-
-    it("no normalisation = no change", () => {
-      const editorState = EditorState.createWithContent(
-        ContentState.createFromBlockArray(
-          convertFromHTML(`<ul><li>UL</li></ul><p>P</p><h2>H2</h2>`),
-        ),
-      )
-      expect(
-        resetBlockType(editorState, [
-          UNSTYLED,
-          UNORDERED_LIST_ITEM,
-          HEADER_TWO,
-        ]),
-      ).toBe(editorState)
-    })
-  })
-
-  describe("#filterInlineStyle", () => {
-    it("works", () => {
-      const contentState = convertFromRaw({
-        entityMap: {},
-        blocks: [
-          {
-            key: "d3071",
-            text: "tes",
-            type: "unstyled",
-            depth: 0,
-            inlineStyleRanges: [
-              {
-                offset: 0,
-                length: 1,
-                style: "BOLD",
-              },
-              {
-                offset: 1,
-                length: 1,
-                style: "ITALIC",
-              },
-              {
-                offset: 2,
-                length: 1,
-                style: "BOLD",
-              },
-              {
-                offset: 2,
-                length: 1,
-                style: "ITALIC",
-              },
-            ],
-            entityRanges: [],
-            data: {},
-          },
-          {
-            key: "abbm4",
-            text: "t",
-            type: "unstyled",
-            depth: 0,
-            inlineStyleRanges: [
-              {
-                offset: 0,
-                length: 1,
-                style: "BOLD",
-              },
-            ],
-            entityRanges: [],
-            data: {},
-          },
-        ],
-      })
-      const editorState = filterInlineStyle(
-        EditorState.createWithContent(contentState),
-        [BOLD],
-      )
-      expect(
-        editorState
-          .getCurrentContent()
-          .getBlockMap()
-          .map((b) => b.getCharacterList().map((c) => c.getStyle()))
-          .toJS(),
-      ).toEqual({ abbm4: [["BOLD"]], d3071: [["BOLD"], [], ["BOLD"]] })
-    })
-  })
-
+describe("entities", () => {
   describe("#resetAtomicBlocks", () => {
     it("works", () => {
       const contentState = convertFromRaw({
@@ -804,29 +458,6 @@ describe("normalize", () => {
             .getEntityAt(0),
         ).not.toBe(null)
       })
-    })
-  })
-
-  describe("#filterEditorState", () => {
-    it("works", () => {
-      const editorState = EditorState.createEmpty()
-      expect(
-        filterEditorState(
-          editorState,
-          1,
-          false,
-          [HEADER_FIVE],
-          [BOLD],
-          [IMAGE],
-        ),
-      ).toBeInstanceOf(EditorState)
-    })
-
-    it(HORIZONTAL_RULE, () => {
-      const editorState = EditorState.createEmpty()
-      expect(
-        filterEditorState(editorState, 1, true, [], [], []),
-      ).toBeInstanceOf(EditorState)
     })
   })
 })
