+++ src/lib/index.test.js
@@ -1,7 +1,18 @@
-import { filterEditorState } from "./index"
+import {
+  preserveAtomicBlocks,
+  resetBlockDepth,
+  resetBlockType,
+  filterInlineStyle,
+  resetAtomicBlocks,
+  filterEntityType,
+  filterEditorState,
+} from "./index"
 
 const pkg = require("../../package.json")
 
+/**
+ * Makes sure the API shape is validated against.
+ */
 describe(pkg.name, () => {
-  it("filterEditorState", expect(filterEditorState).toBeDefined())
+  it("filterEditorState", () => expect(filterEditorState).toBeDefined())
 })
